requirements.txt
click
load-dotenv
plexapi
terminaltables